fixing engine mintbot
